pyproject.toml
[project]
name = "agent-team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.7.6",
    "chromadb>=1.0.15",
    "langchain>=0.3.27",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
]
